requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
